Semantics.Logic.Propositional.Semantics

Description

This is the "internal" module for propositional logic semantics. See the main Semantics.Logic.Propositional module for an explanation of what is going on.

Model Structure

type Interpretation = Map PropositionalVariable Bool

An interpretation for a boolean logic is a mapping from propositional variables to boolean values.

createInterpretation :: [(PropositionalVariable, Bool)] -> Interpretation

Create an interpretation

Evaluation

evaluateIn :: PropositionalFormula -> Interpretation -> Bool

Evaluates a propositional formula in an interpretation.

Let i be an interpretation such that

i 1 -> True
i 2 -> False
i 3 -> True

Then

>>> evaluateIn (Negation (Implication (Disjunction (Atomic 1) (Conjunction (Negation (Atomic 2)) (Atomic 3))) (Atomic 2))) i
True