Semantics.Logic.Propositional.Syntax

Description

This is the "internal" module for propositional logic syntax. See the main Semantics.Logic.Propositional module for an explanation of what is going on.

Abstract Syntax Representation

type PropositionalVariable = Int

A propositional variable is a point.

data PropositionalFormula

A propositional formula is a well formed abstract syntax tree encoding its logical form.

Constructors

Atomic !PropositionalVariable

Negation !PropositionalFormula

Conjunction !PropositionalFormula !PropositionalFormula

Disjunction !PropositionalFormula !PropositionalFormula

Implication !PropositionalFormula !PropositionalFormula