Semantics.Logic.PropositionalModal.Semantics

Documentation

type World = Int

A World is a distinct point with no interal structure.

type Worlds = Set World

Worlds are sets of points (should probably be IntSet).

type AccessibilityRelation = Map World Worlds

An accessibility relation is a map from a point to a set of points.

type KripkeFrame = (Worlds, AccessibilityRelation)

type PropositionsTrueInWorld = Set PropositionalVariable

type ModalInterpretation = Map World PropositionsTrueInWorld

type KripkeModel = (KripkeFrame, ModalInterpretation)

evaluateAtIn :: PropositionalModalFormula -> World -> KripkeModel -> Bool

Evaluates the truth of a formula at a world according to a Kripke model.

Basically a structural recursion over the logical form of the formula. Necessity is evaluated by mapping the evaluation of the formula at each world accessible by the input world, and then folding over that set of booleans with (&&), similarly for possibility but with (||).