Semantics.Logic.PropositionalModal.Semantics

Description

This is the "internal" module for propositional modal logic semantics. See the main Semantics.Logic.PropositionalModal module for an explanation of what is going on.

Model Structure

type World = Int

A World is a distinct point with no interal structure.

type Worlds = Set World

Worlds are sets of points (should probably be IntSet).

type AccessibilityRelation = Map World Worlds

An accessibility relation is a map from a point to a set of points.

type KripkeFrame = (Worlds, AccessibilityRelation)

A Kripke frame is a pair of worlds and an accessibility relation defined on them.

type PropositionsTrueInWorld = Set PropositionalVariable

The set of propositions true in a world.

type ModalInterpretation = Map World PropositionsTrueInWorld

A modal interpretation is a mapping from a world to the set of propositions true in that world.

type KripkeModel = (KripkeFrame, ModalInterpretation)

A Kripke Model is a pair of a Kripke frame and a modal interpretation

Evaluation

evaluateAtIn :: PropositionalModalFormula -> World -> KripkeModel -> Bool

Evaluates the truth of a formula at a world according to a Kripke model.

Let km be a Kripke model with worlds w, w', and w'', an accessibility relation

w  -> w'
w' -> w'
w' -> w''

and a modal interpretation which assigns

1    -> w
1, 2 -> w'
2, 3 -> w''

then

>>> evaluateAtIn (Necessity (Disjunction (Atomic 1) (Atomic 3))) w km
True