Semantics.Logic.Quantificational.Syntax

Description

This is the "internal" module for predicate logic syntax. See the main Semantics.Logic.Quantificational module for an explanation of what is going on.

Abstract Syntax

type QuantificationalTerm = String

A quantificational term is the name of a constant.

data QuantificationalFormula

A quantificational formula is a well formed abstract syntax tree encoding its logical form.

Constructors

Predicate String [QuantificationalTerm]

Equality QuantificationalTerm QuantificationalTerm

Negation QuantificationalFormula

Conjunction QuantificationalFormula QuantificationalFormula

Disjunction QuantificationalFormula QuantificationalFormula

Implication QuantificationalFormula QuantificationalFormula

Universal (QuantificationalTerm -> QuantificationalFormula)

Existential (QuantificationalTerm -> QuantificationalFormula)